requirements.txt
kivy==2.3.0
kivymd==1.2.0
Pillow>=10.0.0
ffpyplayer>=4.5.0
plyer>=2.1.0
